move related user list to group folder
