+++ web/src/admin/groups/GroupViewPage.ts
@@ -1,5 +1,5 @@
 import "@goauthentik/admin/groups/GroupForm";
-import "@goauthentik/admin/users/RelatedUserList";
+import "@goauthentik/app/admin/groups/RelatedUserList";
 import "@goauthentik/app/elements/rbac/ObjectPermissionsPage";
 import { DEFAULT_CONFIG } from "@goauthentik/common/api/config";
 import { EVENT_REFRESH } from "@goauthentik/common/constants";
